requirements.txt
absl-py==0.11.0
aiodns==1.1.1
aiohttp==3.6.2
aioredis==1.3.1
albumentations==0.5.2
alembic==1.5.2
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
atari-py==0.2.6
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
black==19.10b0
bleach==3.1.5
blessings==1.7
blist==1.3.6
box2d-py==2.3.8
branca==0.4.1
cachetools==4.2.1
ccxt==1.33.47
certifi==2020.12.5
cffi==1.14.0
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cliff==3.6.0
cligj==0.7.1
cloudpickle==1.2.1
cmaes==0.8.0
cmd2==1.5.0
colorama==0.4.4
colorful==0.5.4
colorlog==4.7.2
contextvars==2.4
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
descartes==1.1.0
dm-tree==0.1.5
entrypoints==0.3
fastdtw==0.3.4
filelock==3.0.12
Fiona==1.8.13.post1
Flask==1.1.2
folium==0.11.0
fsspec==0.8.5
future==0.18.2
gast==0.3.3
geopandas==0.8.1
google==3.0.0
google-api-core==1.22.2
google-auth==1.25.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
gpustat==0.6.0
GPUtil==1.4.0
grpcio==1.32.0
gym==0.17.2
h11==0.9.0
h5py==2.10.0
hiredis==1.1.0
httptools==0.1.1
idna==2.10
idna-ssl==1.1.0
imageio==2.9.0
imgaug==0.4.0
immutables==0.14
importlib-metadata==3.4.0
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.0
jupyterlab==2.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupyterlmod==2.0.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kornia==0.4.1
llvmlite==0.35.0
lxml==4.4.1
lz4==3.1.0
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
mock==4.0.2
mpmath==1.1.0
msgpack==1.0.0
multidict==4.5.2
munch==2.5.0
nbclient==0.5.1
nbconvert==5.6.1
nbformat==5.0.7
nest-asyncio==1.4.3
networkx==2.5
nose==1.3.7
notebook==6.1.3
numba==0.52.0
numpy==1.19.2
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
omegaconf==2.0.6
opencensus==0.7.10
opencensus-context==0.1.1
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
optuna==2.5.0
osmnx==0.16.2
packaging==20.9
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pathspec==0.8.0
patsy==0.5.1
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
Pillow-SIMD==7.0.0.post3
plotly==4.9.0
pluggy==0.13.1
portalocker==2.2.1
prettytable==0.7.2
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.14.0
psutil==5.7.0
ptyprocess==0.6.0
py==1.9.0
py-spy==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycparser==2.20
pyglet==1.5.0
Pygments==2.6.1
pyparsing==2.4.7
pyperclip==1.7.0
pyproj==2.6.1.post1
pyrsistent==0.16.0
pytest==6.1.2
python-dateutil==2.8.1
python-editor==1.0.4
pytorch-forecasting==0.8.3
pytorch-lightning==1.1.7
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.6
QtPy==1.9.0
ray==0.8.7
redis==3.4.1
regex==2019.11.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7
Rtree==0.9.4
scikit-image==0.17.2
scikit-learn==0.23.0
scipy==1.5.4
seaborn==0.11.0
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.11
stable-baselines==2.10.1
statsmodels==0.11.1
stevedore==3.3.0
stochastic==0.4.0
svgwrite==1.4
sympy==1.6.2
ta==0.7.0
tabulate==0.8.7
tenacity==6.2.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tianshou==0.3.0
tifffile==2020.12.8
timm==0.3.2
toml==0.10.1
torch==1.7.1
torch-encoding==1.2.1
torchvision==0.7.0
tornado==6.0.4
tqdm==4.56.0
traitlets==4.3.3
Tree==0.2.3
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.11.8
uvloop==0.14.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
yarl==1.1.0
zipp==3.4.0
